+++ packages/gamut-styles/src/variables/index.ts
@@ -1,6 +1,7 @@
 export * from './borderRadii';
 export * from './colors';
 export * from './elements';
+export * from './elevation';
 export * from './responsive';
 export * from './spacing';
 export * from './timing';
